Drop `wasmer` in favor of `wasmtime` for tests
Since we've transitioned to `wasmtime` for the entirety of our stack, we
should also use it for the tests in this crate. This commit does exactly
that, doing away with the `wasmer` dependency, and replacing it with
`wasmtime`.

This also also helps in figuring out exactly what's wrong with our test
runs, by eliminating the runtime as a factor - it errors in the same way
in both `debug` and `release` modes.

See-also: #105